trait impl restructure
